Fix GQL 2.0 Schema Resolve Bug
After updating to the latest GQL 2.0 gem we've run into a bug
where dynamic field types such as Unions cannot be resolved
correctly.  This is because the `resolve_type` method on the
schema class returns a two element array / tuple where the first
element is the type and the second element is the value.

This code hooks into this possibility and correctly initializes
the object so that that the `GraphQL::Decorate::TypeAttributes`
can wrap it.